1,769 named EC functions, and the disassembly they are checked against
The annotation layer had 67 rows. It now has 1,769, covering 1,787 of the
2,708 exported functions -- every one of them read from the function's own
disassembly, and every one checked by a second agent that re-read the same
listing and refused the rows it could not support.

1,737 rows were approved by the verifier; 124 were not and did not go in. The
merge refused a further 40 for the mechanical rules: an address that no longer
resolves, a duplicate, a type outside the vocabulary, a comment that named a
register its own listing does not contain. 138 rows are type `unresolved`,
which is a result and not a failure -- a function whose role the bytes do not
determine is described as far as they go and no further.

What a reader gets: `ec/decompiled/<program>/<ADDR>.c` with the function named
instead of `FUN_CODE_…`, a plate comment saying what the code does, and
`ec/decompiled/<program>/<ADDR>.asm` beside it holding the instructions that
said so. The 8051's reserved instruction, the fill bytes, the one-instruction
fall-through fragments that Ghidra's boundaries cut out of straight-line code
-- all named for what they are.

And the whole export is now verified two ways. Every byte of every listing is
the byte in ec/firmware/GMxMGxx_11.800: 45,537 instructions, zero
disagreements. And 45,394 of those 45,537 re-encode to the firmware bytes
through sdas8051, an assembler that never saw it: 99.69%, with no function in
disagreement and 2,574 of 2,705 having every instruction checked.

+++ ghidra/scripts/ApplyAnnotations.java
@@ -77,12 +77,12 @@ public class ApplyAnnotations extends GhidraScript {
                     if (!appliesTo(row[5], programName)) {
                         continue;
                     }
-                    int addr = hex(row[0]);
+                    long addr = hex(row[0]);
                     String name = row[1];
                     Address a = xdata.getAddress(addr);
                     if (a == null) {
                         labelsSkipped++;
-                        problems.add("EXTMEM address out of range: 0x" + Integer.toHexString(addr));
+                        problems.add("EXTMEM address out of range: 0x" + Long.toHexString(addr));
                         continue;
                     }
                     Symbol s = st.getPrimarySymbol(a);
@@ -124,7 +124,7 @@ public class ApplyAnnotations extends GhidraScript {
             if (!mine) {
                 continue;
             }
-            int addr = hex(row[1]);
+            long addr = hex(row[1]);
             String name = row[2];
             String signature = row[3];
             String type = row[4];
@@ -223,8 +223,12 @@ public class ApplyAnnotations extends GhidraScript {
         return dot > 0 ? name.substring(0, dot) : name;
     }
 
-    private static int hex(String s) {
-        return Integer.parseInt(s.trim().replace("0x", ""), 16);
+    private static long hex(String s) {
+        // long, not int, for the same reason as SeedFunctions: a TE image is
+        // loaded at 0xFFF8xxxx on this platform and does not fit a signed
+        // 32-bit int, so an annotation for one of those modules would have
+        // parsed to a negative address and named the wrong function.
+        return Long.parseLong(s.trim().replace("0x", ""), 16);
     }
 
     /**
